[BUGFIX beta] Correctly strip debug statements when used as addon
