+++ lib/javascripts.js
@@ -1,8 +1,5 @@
 /* jshint node:true */
 
-var filterImports = require('babel-plugin-filter-imports');
-var featureFlags  = require('babel-plugin-feature-flags');
-var babel         = require('broccoli-babel-transpiler');
 var merge         = require('broccoli-merge-trees');
 var concat        = require('broccoli-concat');
 var uglify        = require('broccoli-uglify-sourcemap');
@@ -13,84 +10,23 @@ var path          = require('path');
 var Funnel        = require('broccoli-funnel');
 var versionReplace = require('./version-replace');
 var fileCreator   = require('broccoli-file-creator');
-
-function babelOptions(libraryName, _options) {
-  _options = _options || {};
-
-  var options = {
-    whitelist: [
-      'es6.templateLiterals',
-      'es6.parameters',
-      'es6.arrowFunctions',
-      'es6.destructuring',
-      'es6.spread',
-      'es6.properties.computed',
-      'es6.properties.shorthand',
-      'es6.blockScoping',
-      'es6.constants',
-      'es6.modules'
-    ],
-    sourceMaps: false,
-    modules: 'amdStrict',
-    moduleRoot: libraryName,
-    moduleId: true,
-    // Transforms /index.js files to use their containing directory name
-    getModuleId: function (name) {
-      return name.replace(/\/index$/g, '');
-    },
-    resolveModuleSource: function(name, filename) {
-      if (name.indexOf('.') === 0) {
-        return libraryName + '/' + path.join(path.dirname(filename), name);
-      } else {
-        return name;
-      }
-    }
-  };
-
-  Object.keys(_options).forEach(function(opt) {
-    options[opt] = _options[opt];
-  });
-
-  return options;
-}
+var babelBuild    = require('./babel-build');
+var strippedBuild = require('./stripped-build');
 
 function debugBuild(packageName, tree) {
-  var compiled = babel(tree, babelOptions(packageName));
+  var compiled = babelBuild(packageName, tree);
+
   return stew.mv(compiled, packageName);
 }
 
-function strippedBuild(packageName, tree) {
-  var featuresJson = fs.readFileSync('config/features.json', { encoding: 'utf8' });
-  var features = JSON.parse(featuresJson);
-
-  var plugins = [
-    featureFlags({
-      import: { module: 'ember-data/-private/features' },
-      features: features
-    }),
-
-    filterImports({
-      'ember-data/-private/debug': [
-        'assert',
-        'debug',
-        'deprecate',
-        'info',
-        'runInDebug',
-        'warn',
-        'debugSeal'
-      ]
-    })
-  ];
-
+function makeStrippedBuild(packageName, tree) {
   var withoutDebug = new Funnel(tree, {
     exclude: ['ember-data/-private/debug.js']
   });
 
-  var compiled = babel(withoutDebug , babelOptions(packageName, {
-    plugins: plugins
-  }));
+  var stripped = strippedBuild(packageName, withoutDebug);
 
-  return stew.mv(compiled, packageName);
+  return stew.mv(stripped, packageName);
 }
 
 function collapse(tree, outputFileName) {
@@ -130,21 +66,26 @@ function minify(tree) {
   });
 }
 
-
-module.exports = function(tree) {
+function buildEmberInflector() {
   var emberInflector = new Funnel(path.dirname(require.resolve('ember-inflector/addon')), {
     include: ['**/*.js']
   });
+
+  return debugBuild('ember-inflector', emberInflector);
+}
+
+module.exports = function(tree) {
+  var emberInflector = buildEmberInflector();
   var emberData = merge([tree, version()]);
 
   var javascripts = merge([
-    debugBuild('ember-inflector', emberInflector),
+    emberInflector,
     debugBuild('ember-data', emberData)
   ]);
 
   var strippedJavascripts = merge([
-    strippedBuild('ember-inflector', emberInflector),
-    strippedBuild('ember-data', emberData)
+    emberInflector,
+    makeStrippedBuild('ember-data', emberData)
   ]);
 
   var debug = collapse(javascripts, 'ember-data.js');
