+++ package.json
@@ -22,6 +22,9 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "babel-plugin-feature-flags": "^0.2.0",
+    "babel-plugin-filter-imports": "^0.2.0",
+    "broccoli-babel-transpiler": "^5.5.0",
     "broccoli-file-creator": "^1.0.0",
     "broccoli-merge-trees": "^1.0.0",
     "chalk": "^1.1.1",
@@ -35,11 +38,8 @@
     "silent-error": "^1.0.0"
   },
   "devDependencies": {
-    "babel-plugin-feature-flags": "^0.2.0",
-    "babel-plugin-filter-imports": "^0.2.0",
     "bower": "^1.6.5",
     "broccoli-asset-rev": "^2.1.2",
-    "broccoli-babel-transpiler": "^5.5.0",
     "broccoli-concat": "0.0.13",
     "broccoli-funnel": "^1.0.0",
     "broccoli-jscs": "^1.1.0",
